Fix John's complaints
